\name{OPGD-package}
\alias{OPGD-package}
\docType{package}
\title{
  Optimal Projections for Gaussian Discriminants
}
\description{
  Obtains optimal linear projections for Gaussian discriminant analysis and clustering with Gaussian Mixture Models.
}
\details{
  The main functions are OPGD and OPGDcluster, which perform find optimal projections for the classification and clustering problems
respectively.
}
\references{
  Hofmeyr, D., Kamper, F. and Melonas, M. (2020) Optimal Projections for Gaussian Discriminants. \emph{ArXiv preprint}
}
\author{
David P. Hofmeyr and Francois Kamper
}
